Set default value for hostnames, not quite as robust
Needed to do this because sinatra-assetpack creates an instance of
the App class with no arguments to the constructor, I don't know why
it does this but it's something to do with asset compilation. If we
don't provide an default constructor, it blows up and can't compile
the assets ONLY in production mode. Bizarre.